backend/requirements.txt
django
pillow
django-cors-headers
djangorestframework
dj-rest-auth
django-allauth
djoser